+++ docs/misc/changelog.rst
@@ -81,6 +81,7 @@ Documentation:
 - Updated migration guide (@juancroldan)
 - Pinned ``docutils==0.16`` to avoid issue with rtd theme
 - Clarified callback ``save_freq`` definition
+- Remove recurrent policies from ``A2C`` docs (@bstee615)
 
 
 Release 1.0 (2021-03-15)
@@ -697,4 +698,4 @@ And all the contributors:
 @tirafesi @blurLake @koulakis @joeljosephjin @shwang @rk37 @andyshih12 @RaphaelWag @xicocaio
 @diditforlulz273 @liorcohen5 @ManifoldFR @mloo3 @SwamyDev @wmmc88 @megan-klaiber @thisray
 @tfederico @hn2 @LucasAlegre @AptX395 @zampanteymedio @JadenTravnik @decodyng @ardabbour @lorenz-h @mschweizer @lorepieri8 @vwxyzjn
-@ShangqunYu @PierreExeter @JacopoPan @ltbd78 @tom-doerr @Atlis @liusida @09tangriro @amy12xx @juancroldan @benblack769
+@ShangqunYu @PierreExeter @JacopoPan @ltbd78 @tom-doerr @Atlis @liusida @09tangriro @amy12xx @juancroldan @benblack769 @bstee615